# backend-service/doc/swagger.yaml
openapi: 3.0.3
info:
  title: Dell Technologies Resource Management Application Backend API Documentation
  description: >-
    This is the backend API documentation and specifcation for the Dell Technologies Resource Management Web Application, and it is based on the OpenAPI 3.0 specification.


    The API is a RESTful API service which we designed and built to allow us to perform CRUD operations on data stored on a MongoDB Server.


    This application is built using the following technolgoies:

    - [React.js](https://react.dev/)

    - [Node.js](https://nodejs.org/en)

    - [Express.js](https://expressjs.com/)

    - [MongoDB](https://www.mongodb.com/)

    - [Docker](https://www.docker.com/)

    - [Swagger.io](https://swagger.io/)


    This application is built to manage resources much easier 
    and more efficiently helping the organisation to keep 
    track and manage resources more easily. 


    It is built using the MERN stack and it is a full-stack 
    application which can be easily deployed with the help 
    of Docker deployment. We also utilised React.js as a 
    frontend framework which works well with Express.js and 
    Node.js as the backend server. For the database, we have
    used MongoDB, since it is a NoSQL database, it allows 
    for much more flexibility as well as scalability.
    Lastly, for the ease of testing and documentation, we 
    have utilised Swagger.io to document our API endpoints.


...
    below.


...

...
  contact:
...
  license:
...
servers:
...
paths:
...
  /admin/getclient:
    get:
      tags:
        - admin
      summary: Get all client users
      description: Get all client users
      operationId: getAllUsers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "500":
          description: "Error Fetching Clients: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/getclient/{clientId}:
    get:
      tags:
        - admin
      summary: Get a client by ID
      description: Get a client by ID
      operationId: getClientById
      parameters:
        - name: clientId
          in: path
          description: ID of Client to return
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
            application/xml:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Client not found
        "500":
          description: "Error Retrieving Client: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/addclient:
    post:
      tags:
        - admin
      summary: Add a new client user
      description: Add a new client user
      operationId: createClient
      requestBody:
        description: Create a new client user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Client"
          application/xml:
            schema:
              $ref: "#/components/schemas/Client"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Client"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
            application/xml:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: All required fields not filled
        "401":
          description: Authentication failed
        "403":
          description: Access forbidden
        "409":
          description: Username or Email already exists
        "500":
          description: "Error Creating Client: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/updateclient/{clientId}:
    put:
      tags:
        - admin
      summary: Update a client by ID
      description: Update a client by ID
      operationId: updateClient
      parameters:
        - name: clientId
          in: path
          description: ID of Client to update
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      requestBody:
        description: Update an existent client user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Client"
          application/xml:
            schema:
              $ref: "#/components/schemas/Client"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Client"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
            application/xml:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Client not found
        "500":
          description: "Error Updating Client: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/deleteclient/{clientId}:
    delete:
      tags:
        - admin
      summary: Delete a client by ID
      description: Delete a client by ID
      operationId: deleteClient
      parameters:
        - name: clientId
          in: path
          description: ID of Client to delete
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      responses:
        "200":
          description: successful operation
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Client not found
        "500":
          description: "Error Deleting Client: Internal Server Error"
      security:
        - admin_auth_token: []
...
  /client/login:
    get:
      tags:
        - client
      summary: Logs client into the system
      description: Logs client into the system
      operationId: clientLogin
      parameters:
        - name: credential
          in: query
          description: The username / email for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "authentication: successful operation"
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
            application/xml:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: usernmae or password incorrect
        "404":
          description: Invalid username/password supplied
        "500":
          description: "Error Logging In: Internal Server Error"
  /client/logout:
    get:
      tags:
        - client
      summary: Logs out current logged in client session
      description: Logs out current logged in client session
      operationId: clientLogout
      responses:
        "200":
          description: successful operation
        "500":
          description: "Error Logging Out: Internal Server Error"
  /client/createClient:
    post:
      tags:
        - client
      summary: Create a new client user
      description: Create a new client user
      operationId: createUser
      requestBody:
        description: Create a new client user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Client"
          application/xml:
            schema:
              $ref: "#/components/schemas/Client"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Client"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
            application/xml:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        "500":
          description: "Error Creating Client: Internal Server Error"
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        _id:
          type: string
          example: 668d5ba4a5f8eca9ff47283e
          description: MongoDB ObjectID as a hexadecimal string.
        username:
          type: string
          example: theClient
        firstName:
          type: string
          example: Lewis
        lastName:
          type: string
          example: Hamilton
        email:
          type: string
          example: lewis@example.com
        password:
          type: string
          example: "12345"
          description: Password is hashed before storing in the database.
        role:
          type: string
          example: client
          enum:
            - admin
            - trainer
            - client
      xml:
        name: client
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: "##default"
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    admin_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    trainer_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    client_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header